Separate the tests in two folders - unit and integration. Modified the package json scripts
