+++ test/unit/.eslintrc
@@ -1,6 +1,6 @@
 {
   "extends": [
-    "../.eslintrc"
+    "../../.eslintrc"
   ],
   "env": {
     "mocha": true
