+++ test/unit/datastore/filestore.test.js
