+++ test/unit/datastore/syncstore.test.js
