+++ test/unit/entity/metadata.test.js
