+++ test/unit/errors/activeUser.test.js
