+++ test/unit/query.test.js
